Remove category mappings feature, replaced by transaction rules
Category mappings (the category_mappings table and all CRUD) have been
fully replaced by agent-driven transaction rules for categorization.

- Drop category_mappings table via migration
- Remove sync engine mapping fallback (rules → uncategorized now)
- Remove 6 MCP mapping tools and list_unmapped_categories
- Remove REST API mapping endpoints
- Remove admin Mappings and Bulk Edit tabs from categories page
- Simplify CSV import to use uncategorized (rules handle on next sync)
- Update ResetTransactionCategory to set uncategorized instead of
  re-resolving through mappings
- Clean up MergeCategories to skip mapping reassignment
- Remove ~2,800 lines of mapping code, tests, and templates
