+++ internal/api/categories.go
@@ -23,18 +23,6 @@ func ListCategoriesHandler(svc *service.Service) http.HandlerFunc {
 	}
 }
 
-// ListUnmappedCategoriesHandler returns provider categories that have no mapping.
-func ListUnmappedCategoriesHandler(svc *service.Service) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		unmapped, err := svc.ListUnmappedCategories(r.Context())
-		if err != nil {
-			mw.WriteError(w, http.StatusInternalServerError, "INTERNAL_ERROR", "Failed to list unmapped categories")
-			return
-		}
-		writeData(w, unmapped)
-	}
-}
-
 // GetCategoryHandler returns a single category by ID.
 func GetCategoryHandler(svc *service.Service) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
