+++ internal/api/router.go
@@ -50,11 +50,8 @@ func NewRouter(a *app.App, version string) http.Handler {
 		r.Get("/transactions/merchants", MerchantSummaryHandler(svc))
 		r.Get("/transactions/{id}", GetTransactionHandler(svc))
 		r.Get("/categories", ListCategoriesHandler(svc))
-		r.Get("/categories/unmapped", ListUnmappedCategoriesHandler(svc))
 		r.Get("/categories/export", ExportCategoriesTSVHandler(svc))
 		r.Get("/categories/{id}", GetCategoryHandler(svc))
-		r.Get("/category-mappings", ListMappingsHandler(svc))
-		r.Get("/category-mappings/export", ExportMappingsTSVHandler(svc))
 		r.Get("/users", ListUsersHandler(svc))
 		r.Get("/connections", ListConnectionsHandler(svc))
 		r.Get("/connections/{id}/status", GetConnectionStatusHandler(svc))
@@ -81,9 +78,6 @@ func NewRouter(a *app.App, version string) http.Handler {
 			r.Put("/categories/{id}", UpdateCategoryHandler(svc))
 			r.Delete("/categories/{id}", DeleteCategoryHandler(svc))
 			r.Post("/categories/{id}/merge", MergeCategoriesHandler(svc))
-			r.Put("/category-mappings", BulkUpsertMappingsHandler(svc))
-			r.Post("/category-mappings/import", ImportMappingsTSVHandler(svc))
-			r.Delete("/category-mappings/{id}", DeleteMappingHandler(svc))
 			r.Post("/sync", TriggerSyncHandler(svc))
 			r.Post("/transactions/{transaction_id}/comments", CreateCommentHandler(svc))
 			r.Put("/transactions/{transaction_id}/comments/{id}", UpdateCommentHandler(svc))
