+++ internal/service/categories.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"breadbox/internal/db"
-	bsync "breadbox/internal/sync"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
@@ -35,30 +34,10 @@ type CategoryResponse struct {
 	UpdatedAt          string              `json:"updated_at"`
 }
 
-// CategoryMappingResponse represents a category mapping for API responses.
-type CategoryMappingResponse struct {
-	ID                  string `json:"id"`
-	Provider            string `json:"provider"`
-	ProviderCategory    string `json:"provider_category"`
-	CategoryID          string `json:"category_id"`
-	CategorySlug        string `json:"category_slug"`
-	CategoryDisplayName string `json:"category_display_name"`
-	CreatedAt           string `json:"created_at"`
-	UpdatedAt           string `json:"updated_at"`
-}
-
-// UnmappedCategoryPair represents a provider category pair with no mapping.
-type UnmappedCategoryPair struct {
-	Provider string  `json:"provider"`
-	Primary  *string `json:"primary"`
-	Detailed *string `json:"detailed"`
-}
-
 var (
-	ErrCategoryNotFound   = errors.New("category not found")
+	ErrCategoryNotFound    = errors.New("category not found")
 	ErrCategoryUndeletable = errors.New("the uncategorized category cannot be deleted")
-	ErrSlugConflict       = errors.New("category slug already exists")
-	ErrMappingNotFound    = errors.New("category mapping not found")
+	ErrSlugConflict        = errors.New("category slug already exists")
 )
 
 // ListCategories returns all categories as a flat list with parent info.
@@ -385,12 +364,6 @@ func (s *Service) MergeCategories(ctx context.Context, sourceID, targetID string
 		}); err != nil {
 			return fmt.Errorf("reassign child transactions: %w", err)
 		}
-		if err := qtx.ReassignMappingsCategory(ctx, db.ReassignMappingsCategoryParams{
-			CategoryID:   childUID,
-			CategoryID_2: tgtUID,
-		}); err != nil {
-			return fmt.Errorf("reassign child mappings: %w", err)
-		}
 		if err := qtx.ReassignRulesCategory(ctx, db.ReassignRulesCategoryParams{
 			CategoryID:   childUID,
 			CategoryID_2: tgtUID,
@@ -404,13 +377,6 @@ func (s *Service) MergeCategories(ctx context.Context, sourceID, targetID string
 		CategoryID_2: tgtUID,
 	}); err != nil {
 		return fmt.Errorf("reassign transactions: %w", err)
-	}
-
-	if err := qtx.ReassignMappingsCategory(ctx, db.ReassignMappingsCategoryParams{
-		CategoryID:   srcUID,
-		CategoryID_2: tgtUID,
-	}); err != nil {
-		return fmt.Errorf("reassign mappings: %w", err)
 	}
 
 	// Reassign transaction rules so they are re-pointed to the target category
@@ -462,7 +428,8 @@ func (s *Service) SetTransactionCategory(ctx context.Context, txnID, categoryID 
 	return nil
 }
 
-// ResetTransactionCategory clears the manual override and re-resolves category from mappings.
+// ResetTransactionCategory clears the manual override and sets the category to uncategorized.
+// Transaction rules will re-categorize it on the next sync if a matching rule exists.
 func (s *Service) ResetTransactionCategory(ctx context.Context, txnID string) error {
 	txnUID, err := parseUUID(txnID)
 	if err != nil {
@@ -478,35 +445,12 @@ func (s *Service) ResetTransactionCategory(ctx context.Context, txnID string) er
 		return ErrNotFound
 	}
 
-	// Re-resolve: look up the transaction's raw categories and resolve through mappings
-	txn, err := s.Queries.GetTransaction(ctx, txnUID)
+	// Set to uncategorized — rules will re-categorize on next sync.
+	uncategorized, err := s.Queries.GetCategoryBySlug(ctx, "uncategorized")
 	if err != nil {
-		return fmt.Errorf("get transaction: %w", err)
+		return fmt.Errorf("get uncategorized: %w", err)
 	}
-
-	// Determine the provider for this transaction
-	var providerStr string
-	err = s.Pool.QueryRow(ctx,
-		"SELECT c.provider FROM bank_connections c JOIN accounts a ON a.connection_id = c.id WHERE a.id = $1",
-		txn.AccountID).Scan(&providerStr)
-	if err != nil {
-		// If we can't determine provider, set to uncategorized
-		uncategorized, err2 := s.Queries.GetCategoryBySlug(ctx, "uncategorized")
-		if err2 != nil {
-			return fmt.Errorf("get uncategorized: %w", err2)
-		}
-		_, err2 = s.Pool.Exec(ctx, "UPDATE transactions SET category_id = $1 WHERE id = $2", uncategorized.ID, txnUID)
-		return err2
-	}
-
-	// Load mappings for this provider and resolve
-	resolver, err := bsync.NewCategoryResolver(ctx, s.Pool, providerStr)
-	if err != nil {
-		return fmt.Errorf("load resolver: %w", err)
-	}
-
-	categoryID := resolver.Resolve(providerStr, textPtr(txn.CategoryDetailed), textPtr(txn.CategoryPrimary))
-	_, err = s.Pool.Exec(ctx, "UPDATE transactions SET category_id = $1 WHERE id = $2", categoryID, txnUID)
+	_, err = s.Pool.Exec(ctx, "UPDATE transactions SET category_id = $1 WHERE id = $2", uncategorized.ID, txnUID)
 	return err
 }
 
@@ -683,384 +627,6 @@ func (s *Service) BulkRecategorizeByFilter(ctx context.Context, params BulkRecat
 	}, nil
 }
 
-// BulkReMap updates non-overridden transactions from oldCategoryID to newCategoryID
-// where the raw provider category matches.
-func (s *Service) BulkReMap(ctx context.Context, oldCategoryID, newCategoryID string, providerCategory *string) (int64, error) {
-	oldUID, err := parseUUID(oldCategoryID)
-	if err != nil {
-		return 0, ErrCategoryNotFound
-	}
-	newUID, err := parseUUID(newCategoryID)
-	if err != nil {
-		return 0, ErrCategoryNotFound
-	}
-
-	query := "UPDATE transactions SET category_id = $1, updated_at = NOW() WHERE category_id = $2 AND category_override = FALSE AND deleted_at IS NULL"
-	args := []any{newUID, oldUID}
-
-	if providerCategory != nil {
-		query += " AND category_detailed = $3"
-		args = append(args, *providerCategory)
-	}
-
-	tag, err := s.Pool.Exec(ctx, query, args...)
-	if err != nil {
-		return 0, fmt.Errorf("bulk remap: %w", err)
-	}
-	return tag.RowsAffected(), nil
-}
-
-// Category mapping CRUD
-
-// ListMappings returns all category mappings, optionally filtered by provider and/or category slug.
-func (s *Service) ListMappings(ctx context.Context, provider *string, categorySlug ...string) ([]CategoryMappingResponse, error) {
-	var rows []db.ListCategoryMappingsRow
-	var err error
-
-	if provider != nil {
-		typedRows, err2 := s.Queries.ListCategoryMappingsByProvider(ctx, db.ProviderType(*provider))
-		if err2 != nil {
-			return nil, fmt.Errorf("list mappings by provider: %w", err2)
-		}
-		// Convert to common row type
-		for _, r := range typedRows {
-			rows = append(rows, db.ListCategoryMappingsRow{
-				ID:                  r.ID,
-				Provider:            r.Provider,
-				ProviderCategory:    r.ProviderCategory,
-				CategoryID:          r.CategoryID,
-				CreatedAt:           r.CreatedAt,
-				UpdatedAt:           r.UpdatedAt,
-				CategorySlug:        r.CategorySlug,
-				CategoryDisplayName: r.CategoryDisplayName,
-			})
-		}
-	} else {
-		rows, err = s.Queries.ListCategoryMappings(ctx)
-		if err != nil {
-			return nil, fmt.Errorf("list mappings: %w", err)
-		}
-	}
-
-	// Determine optional category slug filter.
-	var slugFilter string
-	if len(categorySlug) > 0 && categorySlug[0] != "" {
-		slugFilter = categorySlug[0]
-	}
-
-	var result []CategoryMappingResponse
-	for _, r := range rows {
-		if slugFilter != "" && r.CategorySlug != slugFilter {
-			continue
-		}
-		result = append(result, CategoryMappingResponse{
-			ID:                  formatUUID(r.ID),
-			Provider:            string(r.Provider),
-			ProviderCategory:    r.ProviderCategory,
-			CategoryID:          formatUUID(r.CategoryID),
-			CategorySlug:        r.CategorySlug,
-			CategoryDisplayName: r.CategoryDisplayName,
-			CreatedAt:           r.CreatedAt.Time.UTC().Format(time.RFC3339),
-			UpdatedAt:           r.UpdatedAt.Time.UTC().Format(time.RFC3339),
-		})
-	}
-	return result, nil
-}
-
-// CreateMapping creates a new category mapping.
-func (s *Service) CreateMapping(ctx context.Context, provider, providerCategory, categoryID string) (*CategoryMappingResponse, error) {
-	catUID, err := parseUUID(categoryID)
-	if err != nil {
-		return nil, ErrCategoryNotFound
-	}
-
-	m, err := s.Queries.InsertCategoryMapping(ctx, db.InsertCategoryMappingParams{
-		Provider:         db.ProviderType(provider),
-		ProviderCategory: providerCategory,
-		CategoryID:       catUID,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("insert mapping: %w", err)
-	}
-
-	// Re-resolve uncategorized transactions that match this new mapping
-	if err := s.reResolveAfterMapping(ctx, provider, providerCategory, catUID); err != nil {
-		return nil, fmt.Errorf("re-resolve transactions: %w", err)
-	}
-
-	// Re-fetch with joined category info
-	return s.getMappingResponse(ctx, m.ID)
-}
-
-// UpdateMapping updates a category mapping's target category.
-func (s *Service) UpdateMapping(ctx context.Context, id, categoryID string) (*CategoryMappingResponse, error) {
-	mUID, err := parseUUID(id)
-	if err != nil {
-		return nil, ErrMappingNotFound
-	}
-	catUID, err := parseUUID(categoryID)
-	if err != nil {
-		return nil, ErrCategoryNotFound
-	}
-
-	// Fetch existing mapping to get provider + provider_category for re-resolution
-	existing, err := s.Queries.GetCategoryMappingByID(ctx, mUID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, ErrMappingNotFound
-		}
-		return nil, fmt.Errorf("get existing mapping: %w", err)
-	}
-
-	m, err := s.Queries.UpdateCategoryMapping(ctx, db.UpdateCategoryMappingParams{
-		ID:         mUID,
-		CategoryID: catUID,
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, ErrMappingNotFound
-		}
-		return nil, fmt.Errorf("update mapping: %w", err)
-	}
-
-	// Re-resolve uncategorized transactions that match this mapping
-	if err := s.reResolveAfterMapping(ctx, string(existing.Provider), existing.ProviderCategory, catUID); err != nil {
-		return nil, fmt.Errorf("re-resolve transactions: %w", err)
-	}
-
-	return s.getMappingResponse(ctx, m.ID)
-}
-
-// DeleteMapping deletes a category mapping.
-func (s *Service) DeleteMapping(ctx context.Context, id string) error {
-	mUID, err := parseUUID(id)
-	if err != nil {
-		return ErrMappingNotFound
-	}
-	return s.Queries.DeleteCategoryMapping(ctx, mUID)
-}
-
-// BulkUpsertMappings upserts multiple mappings at once.
-type BulkMappingEntry struct {
-	Provider         string `json:"provider"`
-	ProviderCategory string `json:"provider_category"`
-	CategoryID       string `json:"category_id"`
-}
-
-func (s *Service) BulkUpsertMappings(ctx context.Context, entries []BulkMappingEntry) (int, error) {
-	count := 0
-	for _, e := range entries {
-		catUID, err := parseUUID(e.CategoryID)
-		if err != nil {
-			return count, fmt.Errorf("invalid category id %s: %w", e.CategoryID, err)
-		}
-		_, err = s.Queries.UpsertCategoryMapping(ctx, db.UpsertCategoryMappingParams{
-			Provider:         db.ProviderType(e.Provider),
-			ProviderCategory: e.ProviderCategory,
-			CategoryID:       catUID,
-		})
-		if err != nil {
-			return count, fmt.Errorf("upsert mapping %s/%s: %w", e.Provider, e.ProviderCategory, err)
-		}
-
-		// Re-resolve uncategorized transactions that match this mapping
-		if err := s.reResolveAfterMapping(ctx, e.Provider, e.ProviderCategory, catUID); err != nil {
-			return count, fmt.Errorf("re-resolve after upsert %s/%s: %w", e.Provider, e.ProviderCategory, err)
-		}
-		count++
-	}
-	return count, nil
-}
-
-// ListUnmappedCategories returns distinct raw category pairs from transactions that resolved to uncategorized.
-func (s *Service) ListUnmappedCategories(ctx context.Context) ([]UnmappedCategoryPair, error) {
-	rows, err := s.Queries.ListUnmappedCategories(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("list unmapped categories: %w", err)
-	}
-
-	var result []UnmappedCategoryPair
-	for _, r := range rows {
-		result = append(result, UnmappedCategoryPair{
-			Provider: string(r.Provider),
-			Primary:  textPtr(r.CategoryPrimary),
-			Detailed: textPtr(r.CategoryDetailed),
-		})
-	}
-	return result, nil
-}
-
-// reResolveAfterMapping updates uncategorized transactions whose raw category
-// strings match the given providerCategory. This ensures that when a mapping is
-// created or updated, existing transactions are immediately re-categorized
-// instead of staying in "uncategorized" until the next sync.
-func (s *Service) reResolveAfterMapping(ctx context.Context, provider, providerCategory string, categoryID pgtype.UUID) error {
-	_, err := s.Pool.Exec(ctx, `
-		UPDATE transactions t
-		SET category_id = $1, updated_at = NOW()
-		FROM accounts a
-		JOIN bank_connections bc ON a.connection_id = bc.id
-		WHERE t.account_id = a.id
-		  AND bc.provider = $2
-		  AND t.category_id = (SELECT id FROM categories WHERE slug = 'uncategorized')
-		  AND t.category_override = FALSE
-		  AND t.deleted_at IS NULL
-		  AND (t.category_detailed = $3 OR t.category_primary = $3)
-	`, categoryID, provider, providerCategory)
-	return err
-}
-
-// getMappingResponse fetches a mapping with its joined category info.
-func (s *Service) getMappingResponse(ctx context.Context, id pgtype.UUID) (*CategoryMappingResponse, error) {
-	m, err := s.Queries.GetCategoryMappingByID(ctx, id)
-	if err != nil {
-		return nil, fmt.Errorf("get mapping: %w", err)
-	}
-
-	cat, err := s.Queries.GetCategoryByID(ctx, m.CategoryID)
-	if err != nil {
-		return nil, fmt.Errorf("get category for mapping: %w", err)
-	}
-
-	return &CategoryMappingResponse{
-		ID:                  formatUUID(m.ID),
-		Provider:            string(m.Provider),
-		ProviderCategory:    m.ProviderCategory,
-		CategoryID:          formatUUID(m.CategoryID),
-		CategorySlug:        cat.Slug,
-		CategoryDisplayName: cat.DisplayName,
-		CreatedAt:           m.CreatedAt.Time.UTC().Format(time.RFC3339),
-		UpdatedAt:           m.UpdatedAt.Time.UTC().Format(time.RFC3339),
-	}, nil
-}
-
-// CreateMappingBySlug creates a category mapping using category slug instead of ID.
-func (s *Service) CreateMappingBySlug(ctx context.Context, provider, providerCategory, categorySlug string) (*CategoryMappingResponse, error) {
-	cat, err := s.Queries.GetCategoryBySlug(ctx, categorySlug)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, fmt.Errorf("category '%s' not found. Use list_categories to see valid slugs", categorySlug)
-		}
-		return nil, fmt.Errorf("lookup category: %w", err)
-	}
-
-	m, err := s.Queries.InsertCategoryMapping(ctx, db.InsertCategoryMappingParams{
-		Provider:         db.ProviderType(provider),
-		ProviderCategory: providerCategory,
-		CategoryID:       cat.ID,
-	})
-	if err != nil {
-		if strings.Contains(err.Error(), "duplicate key") || strings.Contains(err.Error(), "unique constraint") {
-			return nil, fmt.Errorf("mapping already exists for (%s, %s). Use update_category_mapping to change it", provider, providerCategory)
-		}
-		return nil, fmt.Errorf("insert mapping: %w", err)
-	}
-
-	// Re-resolve uncategorized transactions that match this new mapping
-	if err := s.reResolveAfterMapping(ctx, provider, providerCategory, cat.ID); err != nil {
-		return nil, fmt.Errorf("re-resolve transactions: %w", err)
-	}
-
-	return s.getMappingResponse(ctx, m.ID)
-}
-
-// UpdateMappingBySlug updates a category mapping's target using slug, found by ID or (provider, provider_category).
-func (s *Service) UpdateMappingBySlug(ctx context.Context, id *string, provider *string, providerCategory *string, categorySlug string) (*CategoryMappingResponse, error) {
-	cat, err := s.Queries.GetCategoryBySlug(ctx, categorySlug)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, fmt.Errorf("category '%s' not found. Use list_categories to see valid slugs", categorySlug)
-		}
-		return nil, fmt.Errorf("lookup category: %w", err)
-	}
-
-	var mappingID pgtype.UUID
-	if id != nil && *id != "" {
-		mappingID, err = parseUUID(*id)
-		if err != nil {
-			return nil, ErrMappingNotFound
-		}
-	} else if provider != nil && providerCategory != nil {
-		m, err2 := s.Queries.GetCategoryMappingByProviderCategory(ctx, db.GetCategoryMappingByProviderCategoryParams{
-			Provider:         db.ProviderType(*provider),
-			ProviderCategory: *providerCategory,
-		})
-		if err2 != nil {
-			if errors.Is(err2, pgx.ErrNoRows) {
-				return nil, ErrMappingNotFound
-			}
-			return nil, fmt.Errorf("lookup mapping: %w", err2)
-		}
-		mappingID = m.ID
-	} else {
-		return nil, fmt.Errorf("either id or (provider, provider_category) is required")
-	}
-
-	m, err := s.Queries.UpdateCategoryMapping(ctx, db.UpdateCategoryMappingParams{
-		ID:         mappingID,
-		CategoryID: cat.ID,
-	})
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, ErrMappingNotFound
-		}
-		return nil, fmt.Errorf("update mapping: %w", err)
-	}
-
-	// Re-resolve uncategorized transactions that match this mapping
-	// Need to fetch the mapping to get the provider_category for re-resolution
-	existing, err := s.Queries.GetCategoryMappingByID(ctx, mappingID)
-	if err == nil {
-		if err2 := s.reResolveAfterMapping(ctx, string(existing.Provider), existing.ProviderCategory, cat.ID); err2 != nil {
-			return nil, fmt.Errorf("re-resolve transactions: %w", err2)
-		}
-	}
-
-	return s.getMappingResponse(ctx, m.ID)
-}
-
-// DeleteMappingByLookup deletes a mapping by ID or (provider, provider_category).
-func (s *Service) DeleteMappingByLookup(ctx context.Context, id *string, provider *string, providerCategory *string) (string, string, error) {
-	if id != nil && *id != "" {
-		mUID, err := parseUUID(*id)
-		if err != nil {
-			return "", "", ErrMappingNotFound
-		}
-		m, err := s.Queries.GetCategoryMappingByID(ctx, mUID)
-		if err != nil {
-			if errors.Is(err, pgx.ErrNoRows) {
-				return "", "", ErrMappingNotFound
-			}
-			return "", "", fmt.Errorf("get mapping: %w", err)
-		}
-		if err := s.Queries.DeleteCategoryMapping(ctx, mUID); err != nil {
-			return "", "", fmt.Errorf("delete mapping: %w", err)
-		}
-		return string(m.Provider), m.ProviderCategory, nil
-	}
-
-	if provider != nil && providerCategory != nil {
-		_, err := s.Queries.GetCategoryMappingByProviderCategory(ctx, db.GetCategoryMappingByProviderCategoryParams{
-			Provider:         db.ProviderType(*provider),
-			ProviderCategory: *providerCategory,
-		})
-		if err != nil {
-			if errors.Is(err, pgx.ErrNoRows) {
-				return "", "", ErrMappingNotFound
-			}
-			return "", "", fmt.Errorf("lookup mapping: %w", err)
-		}
-		if err := s.Queries.DeleteCategoryMappingByProviderCategory(ctx, db.DeleteCategoryMappingByProviderCategoryParams{
-			Provider:         db.ProviderType(*provider),
-			ProviderCategory: *providerCategory,
-		}); err != nil {
-			return "", "", fmt.Errorf("delete mapping: %w", err)
-		}
-		return *provider, *providerCategory, nil
-	}
-
-	return "", "", fmt.Errorf("either id or (provider, provider_category) is required")
-}
 
 // GenerateSlug creates a URL-safe slug from a display name.
 func GenerateSlug(displayName string) string {
@@ -1100,16 +666,6 @@ type CategoryImportResult struct {
 	Merged    int      `json:"merged"`
 	Deleted   int      `json:"deleted"`
 	Errors    []string `json:"errors,omitempty"`
-}
-
-// MappingImportResult summarizes the outcome of a mapping TSV import.
-type MappingImportResult struct {
-	Created             int      `json:"created"`
-	Updated             int      `json:"updated"`
-	Unchanged           int      `json:"unchanged"`
-	Deleted             int      `json:"deleted"`
-	TransactionsUpdated int64    `json:"transactions_updated"`
-	Errors              []string `json:"errors,omitempty"`
 }
 
 var slugRegexp = regexp.MustCompile(`^[a-z0-9_]+$`)
@@ -1405,178 +961,6 @@ func (s *Service) ImportCategoriesTSV(ctx context.Context, content string, repla
 	return result, nil
 }
 
-// ExportMappingsTSV returns all category mappings as a TSV string.
-func (s *Service) ExportMappingsTSV(ctx context.Context) (string, error) {
-	mappings, err := s.ListMappings(ctx, nil)
-	if err != nil {
-		return "", fmt.Errorf("list mappings: %w", err)
-	}
-
-	headers := []string{"provider", "provider_category", "category_slug"}
-	rows := make([][]string, len(mappings))
-	for i, m := range mappings {
-		rows[i] = []string{m.Provider, m.ProviderCategory, m.CategorySlug}
-	}
-
-	return formatTSV(headers, rows), nil
-}
-
-// ImportMappingsTSV parses TSV content and upserts category mappings.
-// If applyRetroactively is true, ALL non-overridden transactions matching
-// the raw category strings are re-categorized (not just uncategorized ones).
-// If replaceMode is true, mappings not present in the import are deleted.
-func (s *Service) ImportMappingsTSV(ctx context.Context, content string, applyRetroactively bool, replaceMode bool) (*MappingImportResult, error) {
-	rows, err := parseTSV(content, 3, 3)
-	if err != nil {
-		return nil, fmt.Errorf("parse TSV: %w", err)
-	}
-
-	result := &MappingImportResult{}
-
-	for i, cols := range rows {
-		lineNum := i + 2
-
-		provider := strings.TrimSpace(cols[0])
-		providerCategory := strings.TrimSpace(cols[1])
-		categorySlug := strings.TrimSpace(cols[2])
-
-		// Validate provider
-		switch provider {
-		case "plaid", "teller", "csv":
-		default:
-			result.Errors = append(result.Errors, fmt.Sprintf("line %d: invalid provider '%s'", lineNum, provider))
-			continue
-		}
-
-		if providerCategory == "" {
-			result.Errors = append(result.Errors, fmt.Sprintf("line %d: provider_category is required", lineNum))
-			continue
-		}
-		if categorySlug == "" {
-			result.Errors = append(result.Errors, fmt.Sprintf("line %d: category_slug is required", lineNum))
-			continue
-		}
-
-		// Resolve category slug to ID
-		cat, err := s.Queries.GetCategoryBySlug(ctx, categorySlug)
-		if err != nil {
-			if errors.Is(err, pgx.ErrNoRows) {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: category slug '%s' not found", lineNum, categorySlug))
-			} else {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: category lookup error: %v", lineNum, err))
-			}
-			continue
-		}
-
-		// Check if mapping already exists and whether it changed
-		existing, err := s.Queries.GetCategoryMappingByProviderCategory(ctx, db.GetCategoryMappingByProviderCategoryParams{
-			Provider:         db.ProviderType(provider),
-			ProviderCategory: providerCategory,
-		})
-		if err != nil && !errors.Is(err, pgx.ErrNoRows) {
-			result.Errors = append(result.Errors, fmt.Sprintf("line %d: lookup error: %v", lineNum, err))
-			continue
-		}
-
-		if err == nil {
-			// Existing mapping — check if category changed
-			if existing.CategoryID == cat.ID {
-				result.Unchanged++
-				continue
-			}
-			// Update
-			_, err = s.Queries.UpdateCategoryMapping(ctx, db.UpdateCategoryMappingParams{
-				ID:         existing.ID,
-				CategoryID: cat.ID,
-			})
-			if err != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: update error: %v", lineNum, err))
-				continue
-			}
-			result.Updated++
-		} else {
-			// New mapping — create
-			_, err = s.Queries.InsertCategoryMapping(ctx, db.InsertCategoryMappingParams{
-				Provider:         db.ProviderType(provider),
-				ProviderCategory: providerCategory,
-				CategoryID:       cat.ID,
-			})
-			if err != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: create error: %v", lineNum, err))
-				continue
-			}
-			result.Created++
-		}
-
-		// Re-resolve transactions
-		if applyRetroactively {
-			affected, err := s.reResolveAllAfterMapping(ctx, provider, providerCategory, cat.ID)
-			if err != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: re-resolve error: %v", lineNum, err))
-				continue
-			}
-			result.TransactionsUpdated += affected
-		} else {
-			if err := s.reResolveAfterMapping(ctx, provider, providerCategory, cat.ID); err != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("line %d: re-resolve error: %v", lineNum, err))
-				continue
-			}
-		}
-	}
-
-	// Replace mode: delete mappings not present in the import.
-	if replaceMode {
-		type mappingKey struct{ provider, category string }
-		importedKeys := make(map[mappingKey]bool)
-		for _, cols := range rows {
-			provider := strings.TrimSpace(cols[0])
-			providerCategory := strings.TrimSpace(cols[1])
-			if provider != "" && providerCategory != "" {
-				importedKeys[mappingKey{provider, providerCategory}] = true
-			}
-		}
-
-		allMappings, err := s.ListMappings(ctx, nil)
-		if err != nil {
-			result.Errors = append(result.Errors, fmt.Sprintf("replace mode: failed to list mappings: %v", err))
-			return result, nil
-		}
-
-		for _, m := range allMappings {
-			if importedKeys[mappingKey{m.Provider, m.ProviderCategory}] {
-				continue
-			}
-			if err := s.DeleteMapping(ctx, m.ID); err != nil {
-				result.Errors = append(result.Errors, fmt.Sprintf("replace mode: failed to delete mapping %s/%s: %v", m.Provider, m.ProviderCategory, err))
-				continue
-			}
-			result.Deleted++
-		}
-	}
-
-	return result, nil
-}
-
-// reResolveAllAfterMapping updates ALL non-overridden transactions whose raw
-// category strings match the given providerCategory, regardless of their current
-// category_id. This is the retroactive version of reResolveAfterMapping.
-func (s *Service) reResolveAllAfterMapping(ctx context.Context, provider, providerCategory string, categoryID pgtype.UUID) (int64, error) {
-	tag, err := s.Pool.Exec(ctx, `
-		UPDATE transactions t
-		SET category_id = $1, updated_at = NOW()
-		FROM accounts a
-		JOIN bank_connections bc ON a.connection_id = bc.id
-		WHERE t.account_id = a.id
-		  AND bc.provider = $2
-		  AND t.category_override = FALSE
-		  AND t.deleted_at IS NULL
-		  AND (t.category_detailed = $3 OR t.category_primary = $3)
-	`, categoryID, provider, providerCategory)
-	if err != nil {
-		return 0, err
-	}
-	return tag.RowsAffected(), nil
-}
 
 // --- TSV helpers ---
