--- internal/templates/pages/category_mappings.html
@@ -1,151 +1,0 @@
-{{define "content"}}
-{{template "breadcrumb" .}}
-
-<div class="bb-page-header">
-  <div>
-    <h1 class="bb-page-title">Category Mappings</h1>
-    <p class="text-sm text-base-content/50 mt-1">Map provider category strings to your category taxonomy</p>
-  </div>
-  <a href="/categories" class="btn btn-sm btn-ghost rounded-xl gap-2">
-    <i data-lucide="arrow-left" class="w-4 h-4"></i> Categories
-  </a>
-</div>
-
-{{if .UnmappedCategories}}
-<div class="bb-card p-6 mb-6 border-warning/20" x-data="{show: true}" x-show="show">
-  <div class="flex items-center gap-3 mb-4">
-    <div class="flex items-center justify-center w-10 h-10 rounded-xl bg-warning/10">
-      <i data-lucide="alert-triangle" class="w-5 h-5 text-warning"></i>
-    </div>
-    <div>
-      <h3 class="text-sm font-semibold">{{len .UnmappedCategories}} unmapped categories</h3>
-      <p class="text-xs text-base-content/45">These categories from your bank providers need a mapping.</p>
-    </div>
-  </div>
-
-  <div class="overflow-x-auto rounded-xl">
-    <table class="table table-sm">
-      <thead>
-        <tr class="bg-base-200/40">
-          <th class="text-xs font-medium text-base-content/50">Provider</th>
-          <th class="text-xs font-medium text-base-content/50">Primary</th>
-          <th class="text-xs font-medium text-base-content/50">Detailed</th>
-          <th class="text-xs font-medium text-base-content/50">Map To</th>
-        </tr>
-      </thead>
-      <tbody>
-        {{range .UnmappedCategories}}
-        <tr x-data="{categoryId: '', submitting: false}" class="hover:bg-base-200/20 transition-colors duration-150">
-          <td><span class="badge badge-sm {{if eq .Provider "plaid"}}badge-primary{{else if eq .Provider "teller"}}badge-secondary{{else}}badge-accent{{end}} rounded-lg">{{.Provider}}</span></td>
-          <td class="text-sm">{{if .Primary}}{{.Primary}}{{else}}&mdash;{{end}}</td>
-          <td class="text-sm font-mono text-xs">{{if .Detailed}}{{.Detailed}}{{else}}&mdash;{{end}}</td>
-          <td class="flex gap-2 items-center">
-            <select x-model="categoryId" class="select select-sm w-48 bg-base-200/30 border-base-300 rounded-xl">
-              <option value="">Select category...</option>
-              {{range $.AllCategories}}
-              <option value="{{.ID}}">{{if .ParentDisplayName}}{{.ParentDisplayName}} &gt; {{end}}{{.DisplayName}}</option>
-              {{end}}
-            </select>
-            <button class="btn btn-sm btn-primary rounded-xl" :disabled="!categoryId || submitting" @click="
-              submitting = true;
-              let provCat = '{{if .Detailed}}{{.Detailed}}{{else}}{{.Primary}}{{end}}';
-              fetch('/-/category-mappings', {method: 'POST', headers: {'Content-Type': 'application/json'}, body: JSON.stringify({provider: '{{.Provider}}', provider_category: provCat, category_id: categoryId})})
-              .then(r => { if(r.ok) location.reload(); else { submitting = false; window.dispatchEvent(new CustomEvent('bb-toast', {detail:{message:'Failed to create mapping',type:'error'}})); } })
-              .catch(() => { submitting = false; })
-            ">Map</button>
-          </td>
-        </tr>
-        {{end}}
-      </tbody>
-    </table>
-  </div>
-</div>
-{{end}}
-
-<!-- Filter bar -->
-<div class="bb-card px-6 py-4 mb-6">
-  <form method="GET" class="flex items-end gap-4 flex-wrap">
-    <div class="form-control">
-      <label class="label label-text text-xs text-base-content/50 pb-1">Provider</label>
-      <select name="provider" class="select select-bordered select-sm rounded-xl" onchange="this.form.submit()">
-        <option value="">All Providers</option>
-        <option value="plaid"{{if eq .FilterProvider "plaid"}} selected{{end}}>Plaid</option>
-        <option value="teller"{{if eq .FilterProvider "teller"}} selected{{end}}>Teller</option>
-        <option value="csv"{{if eq .FilterProvider "csv"}} selected{{end}}>CSV</option>
-      </select>
-    </div>
-    {{if .FilterProvider}}<a href="/categories/mappings" class="btn btn-sm btn-ghost rounded-xl">Clear</a>{{end}}
-  </form>
-</div>
-
-<!-- Mappings table -->
-<div class="bb-card p-6">
-  <h2 class="text-sm font-semibold text-base-content/70 mb-4">Mappings ({{len .Mappings}})</h2>
-  {{if .Mappings}}
-  <div class="overflow-x-auto rounded-xl">
-    <table class="table table-sm">
-      <thead>
-        <tr class="bg-base-200/40">
-          <th class="text-xs font-medium text-base-content/50">Provider</th>
-          <th class="text-xs font-medium text-base-content/50">Provider Category</th>
-          <th class="text-xs font-medium text-base-content/50">Mapped To</th>
-          <th class="text-xs font-medium text-base-content/50 w-10"></th>
-        </tr>
-      </thead>
-      <tbody>
-        {{range .Mappings}}
-        <tr x-data="{editing: false, categoryId: '{{.CategoryID}}', submitting: false}" class="hover:bg-base-200/20 transition-colors duration-150">
-          <td><span class="badge badge-sm {{if eq .Provider "plaid"}}badge-primary{{else if eq .Provider "teller"}}badge-secondary{{else}}badge-accent{{end}} rounded-lg">{{.Provider}}</span></td>
-          <td class="font-mono text-xs text-base-content/70">{{.ProviderCategory}}</td>
-          <td>
-            <template x-if="!editing">
-              <button class="btn btn-ghost btn-xs gap-2 justify-start rounded-lg" @click="editing = true">
-                <span class="text-sm">{{.CategoryDisplayName}}</span>
-                <i data-lucide="pencil" class="w-3 h-3 opacity-30"></i>
-              </button>
-            </template>
-            <template x-if="editing">
-              <div class="flex gap-2 items-center">
-                <select x-model="categoryId" class="select select-sm w-48 bg-base-200/30 border-base-300 rounded-xl">
-                  {{range $.AllCategories}}
-                  <option value="{{.ID}}">{{if .ParentDisplayName}}{{.ParentDisplayName}} &gt; {{end}}{{.DisplayName}}</option>
-                  {{end}}
-                </select>
-                <button class="btn btn-sm btn-primary rounded-xl" :disabled="submitting" @click="
-                  submitting = true;
-                  fetch('/-/category-mappings/{{.ID}}', {method: 'PUT', headers: {'Content-Type': 'application/json'}, body: JSON.stringify({category_id: categoryId})})
-                  .then(r => { if(r.ok) location.reload(); else { submitting = false; editing = false; } })
-                  .catch(() => { submitting = false; })
-                ">Save</button>
-                <button class="btn btn-sm btn-ghost rounded-xl" @click="editing = false">Cancel</button>
-              </div>
-            </template>
-          </td>
-          <td>
-            <button class="btn btn-ghost btn-xs text-error/50 hover:text-error rounded-lg transition-colors" @click="
-              bbConfirm({title:'Delete mapping?',message:'This category mapping will be removed. Existing transactions are not affected.',confirmLabel:'Delete',variant:'danger'}).then(function(ok){
-                if(!ok) return;
-                fetch('/-/category-mappings/{{.ID}}', {method: 'DELETE'})
-                .then(r => { if(r.ok) location.reload(); })
-              })
-            "><i data-lucide="trash-2" class="w-3.5 h-3.5"></i></button>
-          </td>
-        </tr>
-        {{end}}
-      </tbody>
-    </table>
-  </div>
-  {{else}}
-  <div class="flex flex-col items-center text-center py-12">
-    <div class="w-14 h-14 rounded-xl bg-base-200/60 flex items-center justify-center mb-3">
-      <i data-lucide="arrow-right-left" class="w-7 h-7 text-base-content/30"></i>
-    </div>
-    <p class="text-sm text-base-content/45">No mappings configured yet.</p>
-  </div>
-  {{end}}
-</div>
-
-<script>
-document.addEventListener('DOMContentLoaded', function() { lucide.createIcons(); });
-</script>
-{{end}}
